+++ apps/backend/src/modules/routes/libs/types/types.ts
@@ -1,8 +1,9 @@
 export {
+	type RouteConstructRequestDto,
+	type RouteConstructResponseDto,
+	type RouteCreateRequestDto,
 	type RouteFindAllOptions,
-	type RouteRequestConstructDto,
-	type RouteRequestCreateDto,
-	type RouteRequestPatchDto,
-	type RouteResponseConstructDto,
-	type RouteResponseDto,
+	type RouteGetAllItemResponseDto,
+	type RouteGetByIdResponseDto,
+	type RoutePatchRequestDto,
 } from "@smartscapes/shared";