+++ packages/shared/src/index.ts
@@ -1,4 +1,8 @@
-export { METERS_IN_KM, TIME_UNIT } from "./libs/constants/constants.js";
+export {
+	BYTES_IN_MB,
+	METERS_IN_KM,
+	TIME_UNIT,
+} from "./libs/constants/constants.js";
 export {
 	APIErrorType,
 	APIPath,
@@ -29,6 +33,14 @@ export {
 	type ValueOf,
 } from "./libs/types/types.js";
 export { AuthApiPath, AuthExceptionMessage } from "./modules/auth/auth.js";
+export {
+	FileFolderName,
+	type FileMimeType,
+	FilesApiPath,
+	fileUploadFolderValidationSchema,
+	type FileUploadRequestDto,
+	type FileUploadResponseDto,
+} from "./modules/files/files.js";
 export {
 	GroupExceptionMessage,
 	type GroupItemWithPermissionsDto,
@@ -71,13 +83,17 @@ export {
 	type RouteCategoryGetAllItemResponseDto,
 } from "./modules/route-categories/route-categories.js";
 export {
-	RouteApiPath,
+	type RouteConstructRequestDto,
+	type RouteConstructResponseDto,
+	type RouteCreateRequestDto,
 	type RouteFindAllOptions,
+	type RouteGetAllItemResponseDto,
+	type RouteGetByIdResponseDto,
+	type RoutePatchRequestDto,
 	type RouteRequestConstructDto,
 	type RouteRequestCreateDto,
 	type RouteRequestPatchDto,
-	type RouteResponseConstructDto,
-	type RouteResponseDto,
+	RoutesApiPath,
 	routesConstructValidationSchema,
 	routesCreateValidationSchema,
 	routesSearchQueryValidationSchema,