+++ packages/shared/src/modules/files/libs/types/types.ts
@@ -1,0 +1,3 @@
+export { type FileMimeType } from "./file-mime-type.type.js";
+export { type FileUploadRequestDto } from "./file-upload-request-dto.type.js";
+export { type FileUploadResponseDto } from "./file-upload-response-dto.type.js";