+++ packages/shared/src/modules/routes/libs/enums/route-api-path.enum.ts
@@ -1,5 +1,7 @@
-const RouteApiPath = {
+const RoutesApiPath = {
 	CONSTRUCT: "/construct",
+	ID: "/:id",
+	ROOT: "/",
 } as const;
 
-export { RouteApiPath };
+export { RoutesApiPath };