+++ packages/shared/src/modules/routes/libs/types/route-create-request-dto.type.ts
@@ -1,0 +1,7 @@
+type RouteCreateRequestDto = {
+	description: string;
+	name: string;
+	pois: number[];
+};
+
+export { type RouteCreateRequestDto };