+++ packages/shared/src/modules/routes/libs/types/types.ts
@@ -1,6 +1,10 @@
+export { type RouteConstructRequestDto } from "./route-construct-request-dto.type.js";
+export { type RouteConstructResponseDto } from "./route-construct-response-dto.type.js";
+export { type RouteCreateRequestDto } from "./route-create-request-dto.type.js";
 export { type RouteFindAllOptions } from "./route-find-all-options.type.js";
+export { type RouteGetAllItemResponseDto } from "./route-get-all-item-response-dto.type.js";
+export { type RouteGetByIdResponseDto } from "./route-get-by-id-response-dto.type.js";
+export { type RoutePatchRequestDto } from "./route-patch-request.dto.js";
 export { type RouteRequestConstructDto } from "./route-request-construct-dto.type.js";
 export { type RouteRequestCreateDto } from "./route-request-dto.type.js";
 export { type RouteRequestPatchDto } from "./route-request-update.dto.js";
-export { type RouteResponseConstructDto } from "./route-response-construct-dto.type.js";
-export { type RouteResponseDto } from "./route-response-dto.type.js";